+++ tests/Coverage/CoverageOverlay.test.ts
@@ -23,26 +23,44 @@ jest.mock('vscode', () => {
 import { CoverageOverlay } from '../../src/Coverage/CoverageOverlay'
 import { DefaultFormatter } from '../../src/Coverage/Formatters/DefaultFormatter'
 import { hasDocument } from '../../src/editor'
+import { ICoverageFormatterSettings } from '../../src/Settings'
 
 describe('CoverageOverlay', () => {
   const coverageMapProvider: any = {}
 
+  const settings: ICoverageFormatterSettings = {
+    gutterFormatter: {
+      uncoveredLine: {
+        backgroundColor: '',
+        gutterIconPath: './images/GutterFormatter/uncovered-gutter-icon.svg',
+      },
+      partiallyCoveredLine: {
+        backgroundColor: '',
+        gutterIconPath: './images/GutterFormatter/partially-covered-gutter-icon.svg',
+      },
+      coveredLine: {
+        backgroundColor: '',
+        gutterIconPath: './images/GutterFormatter/covered-gutter-icon.svg',
+      },
+    },
+  }
+
   describe('constructor', () => {
     it('should set the default visibility', () => {
-      const sut = new CoverageOverlay(null, coverageMapProvider)
+      const sut = new CoverageOverlay(null, coverageMapProvider, settings)
 
       expect(sut.enabled).toBe(CoverageOverlay.defaultVisibility)
     })
 
     it('should set the visibility if provided', () => {
       const enabled = !CoverageOverlay.defaultVisibility
-      const sut = new CoverageOverlay(null, coverageMapProvider, enabled)
+      const sut = new CoverageOverlay(null, coverageMapProvider, settings, enabled)
 
       expect(sut.enabled).toBe(enabled)
     })
 
     it('should set the default overlay formatter', () => {
-      const sut = new CoverageOverlay(null, coverageMapProvider)
+      const sut = new CoverageOverlay(null, coverageMapProvider, settings)
 
       expect(DefaultFormatter).toBeCalledWith(coverageMapProvider)
       expect(sut.formatter).toBeInstanceOf(DefaultFormatter)
@@ -53,7 +71,7 @@ describe('CoverageOverlay', () => {
     describe('get', () => {
       it('should return the overlay visibility', () => {
         const expected = true
-        const sut = new CoverageOverlay(null, coverageMapProvider, expected)
+        const sut = new CoverageOverlay(null, coverageMapProvider, settings, expected)
 
         expect(sut.enabled).toBe(expected)
       })
@@ -62,7 +80,7 @@ describe('CoverageOverlay', () => {
     describe('set', () => {
       it('should set the overlay visibility', () => {
         const expected = true
-        const sut = new CoverageOverlay(null, coverageMapProvider, !expected)
+        const sut = new CoverageOverlay(null, coverageMapProvider, settings, !expected)
         sut.updateVisibleEditors = jest.fn()
         sut.enabled = expected
 
@@ -70,7 +88,7 @@ describe('CoverageOverlay', () => {
       })
 
       it('should refresh the overlays in visible editors', () => {
-        const sut = new CoverageOverlay(null, coverageMapProvider)
+        const sut = new CoverageOverlay(null, coverageMapProvider, settings)
         sut.updateVisibleEditors = jest.fn()
         sut.enabled = true
 
@@ -82,7 +100,7 @@ describe('CoverageOverlay', () => {
   describe('toggleVisibility()', () => {
     it('should enable the overlay when disabled', () => {
       const enabled = false
-      const sut = new CoverageOverlay(null, coverageMapProvider, enabled)
+      const sut = new CoverageOverlay(null, coverageMapProvider, settings, enabled)
       sut.updateVisibleEditors = jest.fn()
       sut.toggleVisibility()
 
@@ -91,7 +109,7 @@ describe('CoverageOverlay', () => {
 
     it('should disable the overlay when enabled', () => {
       const enabled = true
-      const sut = new CoverageOverlay(null, coverageMapProvider, enabled)
+      const sut = new CoverageOverlay(null, coverageMapProvider, settings, enabled)
       sut.updateVisibleEditors = jest.fn()
       sut.toggleVisibility()
 
@@ -99,7 +117,7 @@ describe('CoverageOverlay', () => {
     })
 
     it('should refresh the overlays in visible editors', () => {
-      const sut = new CoverageOverlay(null, coverageMapProvider)
+      const sut = new CoverageOverlay(null, coverageMapProvider, settings)
       sut.updateVisibleEditors = jest.fn()
       sut.toggleVisibility()
 
@@ -112,7 +130,7 @@ describe('CoverageOverlay', () => {
       const editors = [{}, {}, {}]
       vscodeProperties.window.visibleTextEditors.mockReturnValueOnce(editors)
 
-      const sut = new CoverageOverlay(null, coverageMapProvider)
+      const sut = new CoverageOverlay(null, coverageMapProvider, settings)
       sut.update = jest.fn()
       sut.updateVisibleEditors()
 
@@ -124,7 +142,7 @@ describe('CoverageOverlay', () => {
 
   describe('update()', () => {
     it('should do nothing if the editor does not have a valid document', () => {
-      const sut = new CoverageOverlay(null, coverageMapProvider)
+      const sut = new CoverageOverlay(null, coverageMapProvider, settings)
       ;(hasDocument as jest.Mock<{}>).mockReturnValueOnce(false)
 
       const editor: any = {}
@@ -136,7 +154,7 @@ describe('CoverageOverlay', () => {
 
     it('should add the overlay when enabled', () => {
       const enabled = true
-      const sut = new CoverageOverlay(null, coverageMapProvider, enabled)
+      const sut = new CoverageOverlay(null, coverageMapProvider, settings, enabled)
       ;(hasDocument as jest.Mock<{}>).mockReturnValueOnce(true)
 
       const editor: any = {}
@@ -147,7 +165,7 @@ describe('CoverageOverlay', () => {
 
     it('should remove the overlay when disabled', () => {
       const enabled = false
-      const sut = new CoverageOverlay(null, coverageMapProvider, enabled)
+      const sut = new CoverageOverlay(null, coverageMapProvider, settings, enabled)
       ;(hasDocument as jest.Mock<{}>).mockReturnValueOnce(true)
 
       const editor: any = {}